chore: Remove duplicate code (#1155)
