+++ app/src/main/res/values/strings.xml
@@ -174,6 +174,7 @@
     <string name="speakers">Speakers</string>
     <string name="sponsors">Sponsors</string>
     <string name="create_sponsor">Create Sponsor</string>
+    <string name="update_sponsor">Update Sponsor</string>
     <string name="sponsor_name">Sponsor Name</string>
     <string name="sponsor_description">Description</string>
     <string name="sponsor_url">URL</string>
